Added payload and user relationships
Having trouble with the sinatra error so we used binding.pry in server.rb. We decided to create a payload connected to the user and ran into the issue of 'no relationship.' Then we added two new migrations with belongs_to/has_many between payloads and users. Next step: finish the payload test and begin creating models/migrating.