Resync'd database and deleted debugging statements after first test in create_payload_test passed
DJ and I were getting an error (NoMethodError: undefined method `val' for #<Arel::Nodes::BindParam) earlier this evening; the database wasnt recognizing the add_column :user_id that we migrated into the test database under the payload table (which we were intending to be a foreign key. We tried debugging but couldnt figure out why we were getting the error. I rolled back the payload migration and re-migrated, which yielded a different error (PG::UndefinedTable: ERROR:  relation 'user_id' does not exist). Stackoverflow advised RAILS_ENV=test rake db:drop db:create db:migrate, so I did that, essentially remigrating everything and then re-ran the test in create payload and the test passed. I deleted all the 'puts' and pry debugging commands from the code. Next step: for some reason the last assertion of the test is not passing so getting that to pass would be the next step!