Added a dashboard.erb file to display the items needed for the app details page
Added skips to all except first test in the user_views_app_details_test file. Wrote the beginnings of an erb dashboard file so that each url in the database will display on the page in order of most to least viewed. Next step: work on building the method to do that in the user model and build the url table.