Added a skip to payload test, added test to user_views_app test, and added line to erb dashboard
skipped payload test which is now failing with the changes to payload table, added a test to user_views_app to show that the user exists and displays on the view page, and added one line to the erb dashboard to make the first test pass. I also reordered the tests minorly, bringing up sad path to second test. Tests wont pass until we have mock data to input.