Changed fixtures to function properly.
Trey's fixture file was setup great, but the payloads weren't being created in the database. I added a user to the database and connected the payloads to that user. The data is now created correctly, but still not accessible. We need to run the file or wrap the content in a module so we can inherit it somewhere. Tests not passing, but not completely broken