Changed first user_test so it would pass
Test was failing because change to the test_helper included a user with id and rootUrl identical to the one we were trying to instantiate in the first test of user_test. Made a very small modification to the identifier and rootUrl so that test_helper and and user test are unique.